// frontend/src/lib/api/generated/models/AnswerItem.ts
/**
 * Roominar
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class AnswerItem {
    'id': number;
    'questionId': number;
    'answer': string;
    'orderNumber': number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "questionId",
            "baseName": "question_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "answer",
            "baseName": "answer",
            "type": "string",
            "format": ""
        },
        {
            "name": "orderNumber",
            "baseName": "order_number",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AnswerItem.attributeTypeMap;
    }

    public constructor() {
    }
}